스폰서 배너 미출력 버그 해결
